% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hf_predict_lines}
\alias{hf_predict_lines}
\title{Predicted Lines for Ribbon Seals}
\format{
\subsection{\code{hf_predict_lines}}{

An sf data frame with 112 rows and 12 columns:
\describe{
\item{speno}{Seal Identifier ('SPEcimen NO.')}
\item{geometry}{Simple Feature LINESTRING geometry with CRS:+proj=laea +lat_0=90 +lon_0=180 +x_0=0 +y_0=0 +datum=WGS84 +units=km +no_defs +type=crs}
\item{species}{seal species; common name}
\item{age}{age class of seal at start of deployment}
\item{sex}{sex of seal. "F"= female; "M"=male}
\item{deploy_dt}{date-time value in UTC that deployment started}
\item{field_effort_lku}{categorical identifier for the field effort}
\item{deployids}{list of deployids; more than one if multiple tags}
\item{serial_nums}{list of serial numbers; more than one if mulitple tags}
\item{ptts}{Argos PTT IDs; more than one if mulitple tags}
\item{tag_families}{tag family; more than one if multiple tags}
\item{end_dt}{date-time value in UTC when the last deployment ended}
}
}
}
\source{
\url{https://github.com/noaa-afsc/mml-cefi-iceseal-data/raw/main/locs_obs/locs_obs.parquet}
}
\usage{
hf_predict_lines
}
\description{
Predicted lines for ribbon seals based on the \code{aniMotum} package
}
\keyword{datasets}
